Restructure discovery module: create utils.py, extract prompt template, fix naming
- Create mlflow/genai/discovery/utils.py with _TokenCounter, _call_llm(),
  build_summary(), and log_discovery_artifacts()
- Move CLUSTER_LABELS_PROMPT_TEMPLATE to constants.py
- Fix single-letter variables in extraction.py (a→assessment, r→rationale)
- Add bidirectional cross-references between session_utils.py and sampling.py
  group_traces_by_session implementations
- Update imports across clustering.py, extraction.py, pipeline.py, and tests
